requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.0
PyPDF2==3.0.1
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# Email
emails==0.6.0

# Validation
pydantic[email]==2.5.0

# Logging
structlog==23.2.0

# Optional: Redis for caching
redis==5.0.1

# Optional: Sentry for error tracking
sentry-sdk[fastapi]==1.38.0